+++ packages/node-core/src/indexer/storeCache/storeCache.service.ts
@@ -79,8 +79,14 @@ export class StoreCacheService extends BaseCacheService {
     if (!this.cachedModels[entity]) {
       const model = this.sequelize.model(entity);
       assert(model, `model ${entity} not exists`);
-      const cacheModel = new CachedModel(model, this._historical, this.config, this._useCockroachDb);
-      cacheModel.init(this.getNextStoreOperationIndex.bind(this));
+      const cacheModel = new CachedModel(
+        model,
+        this._historical,
+        this.config,
+        this.getNextStoreOperationIndex.bind(this),
+        () => this.flushCache(true),
+        this._useCockroachDb
+      );
       this.cachedModels[entity] = cacheModel;
     }
     return this.cachedModels[entity] as unknown as ICachedModel<T>;