+++ packages/types/CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - New LightBlock type (#1968)
+- `getByFields` function to store interface (#1993)
 
 ## [2.2.0] - 2023-08-16
 ### Added